Fix floating X window size preservation when toggling the mode-line
Specifically, the window height is now correctly preserved when
spacious-padding-mode is enabled.

The previous version:

1. Only took the mode-line into account, not any other padding.
2. Used set-frame-height which, in my experience, didn't work reliably.

The new version:

1. Computes the new/old "bottom offset" by comparing the window edges to
the frame edges.
2. Adjusts the window height based on the delta.
3. Uses exwm-layout-enlarge-window to operate on the EXWM window
directly which, in my experience, more reliably sets the exact height.

* exwm-layout.el (exwm-layout--window-bottom-offset): Add a new function
to compute the distance between the bottom of an Emacs window and the
bottom of its frame.
(exwm-layout-hide-mode-line, exwm-layout-show-mode-line): Use the new
function to preserve the X window's height after showing/hiding the
mode-line.