Implement X window icon support
An X window's icon can now be retrieved via calling exwm-icon, if
available.

* exwm-core.el (exwm--icon): Add a variable to cache the window icon. We
store it instead of retrieving it every time because retrieving,
parsing, and turning it into an image we can display is somewhat
expensive.
* exwm.el (exwm--icon-index-colors, exwm--icon-colornames):
(exwm--icon-to-xpm): Add functions to convert X window icons into
XPM images. I chose XPM because it's a relatively simple image format
making it possible to perform this conversion in Emacs without any
external tools.
(exwm--update-icon): Add a function to update the buffer-local
exwm--icon variable when the X window icon changes.
(exwm--on-PropertyNotify): Call this function whenever the icon changes.
(exwm-icon): Add a function to retrieve the X window's current icon as
an Emacs image object.
(exwm--init-icccm-ewmh): Indicate that we support window icons.
(exwm-update-icon-hook): Add a hook to indicate that the window's icon
has been updated.
* exwm-manage.el (exwm-manage--manage-window): Retrieve the X window's
icon when we first manage the window.