Revert "Place floating windows on the correct monitor"
This reverts commit bc2c8eea1fbfc4cc337fcd8607e15055e9f8f715.

The "geometry" is relative to the window's parent. We may need to adjust
the requested geometry relative to the parent (although it SHOULD be the
EXWM root window) , but doing so relative to the screen is definitely
not correct.

I'm not sure why chrome seems to be opening windows on the wrong screen,
but that seems to be a different issue.