Rework closing of "hung" windows
In X, there are 4 ways to "kill":

1. If supported, you can ASK the application to close the window with
   WM_DELETE_WINDOW. This gives the application the chance to, e.g., ask
   the user if they want to save their work first.
2. You can destroy the window directly and immediately.
3. You can kill the X connection associated with the client that owns
   the window. This is what xkill does and generally results in the
   application exiting and/or crashing.
4. You can directly lookup the X client's PID and kill it.

In terms of other window managers:

- i3 does 1 if supported, and 2 if not. It used to fallback on 3, but it
  changed the default behavior at some point.
- BSPWM and DWM do 1 if supported, and 3 if not.

More complex window managers likely implement some kind of timeout
mechanism as EXWM does, but I'm not going to go digging through GNOME's
code.

EXWM, on the other hand does 1 if supported, 2 if 1 isn't supported,
then 4 if 1 is supported but the application isn't responding to pings,
finally falling back on 3 if it can't get the PID of the application
associated with the X window.

This commit changes this to behave more like other window managers:

1. We attempt 1 if supported, 2 if 1 is unsupported.
2. If 1 is supported and we see that the application isn't responded, we
   fallback on 2.

Falling back on 3 in that second case would also be reasonable, but 2 is
the more conservative option.

NOBODY except EXWM does 4 because, as the comment in the removed code
states: programs (potentially running as other users) can simply lie
about their PID.

* exwm-core.el (exwm--ping): Add a variable to track the ping counter on
an X window
* exwm.el (exwm--on-wm-protocols): Update `exwm--ping' when a window
responds to a ping.
* exwm-manage.el (exwm-manage--ping-lock): Remove the ping lock because
it assumes we can only be pinging a single window at a time.
(exwm-manage--ping): Add a new function to ping a window, returning
either nil (when pinging is unsupported) or the current ping
count (before sending this ping).
(exwm-manage--kill-buffer-timeout-function): Add a function to check if
the X window is unresponsive. We use this instead of blocking on process
output.
(exwm-manage--kill-buffer-query-function): Implement the new X window
killing logic described above.
(exwm-manage--kill-client): Simplify `exwm-manage--kill-client' to only
send the KillClient message and never attempt to send a signal to the
underlying process.