Refuse to delete active workspaces when there all are active
Deleting a workspace in this case will attempt to "steal" a workspace
from another monitor. This will fail and we'll end up re-creating a
workspace, causing us to juggle workspaces around.

An alternative would be to ALLOW ourselves to delete the workspace in
this case, but refuse to steal an active workspace to replace it AND
refuse to re-create it. But, IMO, the user likely wants one workspace
per monitor and would rather not delete the last workspace on a monitor
when there are none available to replace it.

* exwm-workspace.el (exwm-workspace-delete): Refuse to delete thej
current workspace if we don't have any to replace it.
(exwm-workspace--get-next-workspace): Make it possible to get the next
workspace, including active workspaces.