Document and warn about XIM deadlocks
Emacs's built-in XIM support can deadlock when trying to communicate
with EXWM. This can be fixed by configuring Emacs to NOT use XIM, so we
now try to detect potential deadlocks and warn about them.

* exwm-xim.el: Document how to configure Emacs to not use XIM.
(exwm-xim--deadlock-warning): Add a function that warns when Emacs may
deadlock with EXWM's XIM module.
(exwm-xim--init): Use it.  (Bug #122)